plan_images_training/requirements_enhanced.txt
# Enhanced Architectural Training System Requirements
# YOLOv8 + CLIP + BLIP Integration

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Transformers for CLIP and BLIP
transformers>=4.30.0
tokenizers>=0.13.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
pathlib2>=2.3.0

# Optional: GPU acceleration
# cuda-python>=12.0.0  # Uncomment if using CUDA

# Optional: Additional features
# accelerate>=0.20.0  # For distributed training
# wandb>=0.15.0       # For experiment tracking 